UI: Replace unneeded GitHub orgs tab with a Threads tab on loops
The threads tab shows an aggregation of threads linked to repositories
of the loop.

Change-type: minor
